Move itoh to FC hex.hpp
