+++ plugins/chain_plugin/include/eosio/chain_plugin/get_info_db.hpp
@@ -61,15 +61,6 @@ namespace eosio::chain_apis {
    private:
       std::unique_ptr<struct get_info_db_impl> _impl;
    }; // get_info_db
-
-   template<typename I>
-   std::string itoh(I n, size_t hlen = sizeof(I)<<1) {
-       static const char* digits = "0123456789abcdef";
-       std::string r(hlen, '0');
-       for(size_t i = 0, j = (hlen - 1) * 4 ; i < hlen; ++i, j -= 4)
-           r[i] = digits[(n>>j) & 0x0f];
-       return r;
-   }
 } // namespace eosio::chain_apis
 
 FC_REFLECT(eosio::chain_apis::get_info_db::get_info_results,
