+++ plugins/prometheus_plugin/metrics.hpp
@@ -4,7 +4,6 @@
 #include <eosio/net_plugin/net_plugin.hpp>
 #include <eosio/producer_plugin/producer_plugin.hpp>
 #include <eosio/chain_plugin/tracked_votes.hpp>
-#include <eosio/chain_plugin/get_info_db.hpp>  // for itoh
 
 #include <prometheus/counter.h>
 #include <prometheus/info.h>
@@ -284,7 +283,7 @@ struct catalog_type {
 
    void update_prometheus_info() {
       info_details = info.Add({
-            {"server_version", chain_apis::itoh(static_cast<uint32_t>(app().version()))},
+            {"server_version", fc::itoh(static_cast<uint32_t>(app().version()))},
             {"chain_id", app().get_plugin<chain_plugin>().get_chain_id()},
             {"server_version_string", app().version_string()},
             {"server_full_version_string", app().full_version_string()},
